Created new file for experiment type hints
